requirements.txt
streamlit
requests
cloudinary
pillow
pytz
google-auth-oauthlib
google-auth-httplib2
google-api-python-client